Time data can be rolled back; Endless questions; improved text UI
If the average percent differences of the average times for the problem set exceeds 2, then the user will have the option to roll back all data accumulated during that round.

User can now choose to have an endless question mode with the set only ending when he/she clicks "finish set." This will allow for future achievements that rely on set length (e.g. complete an easy addition set with over 100 questions correct).

Added methods that allow text to be displayed at the center of a specified rectangle. This makes frontend coding much easier since displaying text now follows a more systematic approach rather than arbitrary constants to determine the location. The method also works for an array of strings to print out multiple lines.

+++ UserData.java
@@ -11,12 +11,12 @@ public class UserData implements Serializable{
 	public int[][] questionsCompleted = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}}; //groups are the question types, inner brackets are the question difficulties
 	public int[][] questionsCorrect = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}}; //groups are the question types, inner brackets are the question difficulties
 
-	public double[][] timeAverageSum = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}};
+	public long[][] timeAverageSum = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}};
 	public int[][] timeAverageCount = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}};
-	public double[][] tempTimeAverageSum = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}};
+	public long[][] tempTimeAverageSum = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}};
 	public int[][] tempTimeAverageCount = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}};
 	
-	public double[][] recordTimeIndividual = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}};
+	public long[][] recordTimeIndividual = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}};
 	
 	public int[][] vanillaAchLevel = {{0, 0, 0, 0},{0, 0, 0, 0},{0, 0, 0, 0},{0, 0, 0, 0}};
