+++ dataUpdater.java
@@ -10,7 +10,6 @@ public class dataUpdater{
 	private String[] difficultyText = {"Easy","Medium","Hard","Insane"};
 	private String[] typeText = {"Addition", "Subtraction", "Multiplication", "Division"}; 
 
-	
 	public void updateQuestionComplete(int type, int difficulty){
 		uData.questionsCompleted[type][difficulty] = uData.questionsCompleted[type][difficulty] + 1;
 	}
@@ -22,9 +21,9 @@ public class dataUpdater{
 		int[][] questionsCompleted = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}}; //groups are the question types, inner brackets are the question difficulties
 		int[][] questionsCorrect = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}}; //groups are the question types, inner brackets are the question difficulties
 		int[][] vanillaAchLevel = {{0, 0, 0, 0},{0, 0, 0, 0},{0, 0, 0, 0},{0, 0, 0, 0}};
-		double[][] timeAverageSum = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}};
+		long[][] timeAverageSum = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}};
 		int[][] timeAverageCount = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}};
-		double[][] recordTimeIndividual = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}};
+		long[][] recordTimeIndividual = {{0,0,0,0}, {0,0,0,0}, {0,0,0,0}, {0,0,0,0}};
 		uData.questionsCompleted = questionsCompleted;
 		uData.questionsCorrect = questionsCorrect;
 		uData.vanillaAchLevel = vanillaAchLevel;
@@ -34,12 +33,13 @@ public class dataUpdater{
 		
 	}
 							
-	public void updateTime(int type, int difficulty, double timeMillis){
+	public void updateTime(int type, int difficulty, long timeMillis){
 		uData.tempTimeAverageSum[type][difficulty] += timeMillis;
 		uData.tempTimeAverageCount[type][difficulty]++;
-		
+			
 		uData.timeAverageSum[type][difficulty] += timeMillis;
 		uData.timeAverageCount[type][difficulty]++;
+		
 		
 		if(uData.recordTimeIndividual[type][difficulty] == 0 || timeMillis < uData.recordTimeIndividual[type][difficulty]) {
 			uData.recordTimeIndividual[type][difficulty] = timeMillis;
